chore(sidecar): move tests
